+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - add `dependabot` support ([#130](https://github.com/JuliaData/XLSX.jl/issues/130))
 - add ability to read native template (`.xltx`) files ([#293](https://github.com/JuliaData/XLSX.jl/issues/293))
 - add `getDefinedNames` to mirror `addDefinedName`
+- support adding png, jpg or gif images into a sheet ([#134](https://github.com/JuliaData/XLSX.jl/issues/134))
 
 ## [v0.11.7](https://github.com/JuliaData/XLSX.jl/tree/v0.11.7) - 2026-05-07
 - Fix issue [#368](https://github.com/JuliaData/XLSX.jl/issues/368) for both reading and writing